requirements.txt
django>=5.0,<5.2
djangorestframework>=3.15,<4.0
django-environ>=0.11,<1.0
django-cors-headers>=4.3,<5.0
djangorestframework-simplejwt>=5.3,<6.0
psycopg2-binary>=2.9,<3.0
